[change] Remove unneeded ssl package implementation definitions
Which ssl implementation is used by the openwisp-monitoring agent is
decided in the curl package. The openwisp-agent is not a compiled
service. Therefore, the division into diffrent package is not necessary.
This commit removes all ssl package versions and the dependency to
openwisp-config-<ssl> (switched to the new openwisp-config without SSL variant).

Signed-off-by: Florian Eckert <fe@dev.tdt.de>